Add normal mode mapping to browser buffer lines in order
The default fzf buffer lines listing `BLines` puts the lines in relevance order
(according to the fzf algorithm).  That is not always what you want, so this is
an alternate version that uses fzf but keeps the lines in source order.

A good use for this is quickly viewing an outline of spec files; filtering by
`describe | it` will show a nice hierarchical view of the specs.