✨ Add bash readline binding to pick any word from tmux pane
Add a keybinding to the sequence `Ctrl-x ;` to interactively pick and insert any
word from the tmux pane history.

The secret sauce here is the `tmux capture-pane` command, which allows us to
programmatically access all the words that have appeared on screen.  After that
we massage the list then pick a word with FZF.

The approach to adding the binding is copied from the fzf bash keybinding configuration.