Only change tmux panes one at a time
The flag to allow repeat was nice to move multiple panes without needing the
prefix again but had the unfortunate side effect of making tmux wait to see what
the next key press was after changing panes.  That meant that `hjkl` following
keys would be caught by tmux as further pane changes, not sent to the selected
pane.