Add alias for git time travel
The pipeline is deliberately split into two, the first using `fzf` to select the
commit hash, the second using `bat` to pretty print the file at that version.
This is because it turns out it's difficult for multiple rich terminal tools to
use a terminal at the same time if they change it's input processing mode.

This was discussed in a `bat` issue:
https://github.com/sharkdp/bat/issues/1650