feat: Add binstubs for dockerenvs
To avoid file system changes, the binstubs for programs inside the docker
environment are created as aliases in the local shell only.  Unfortunately that
isn't possible without using `eval` to execute commands in the parent bash
process.

This follows the example of `ssh-agent` and `docker-machine` which provide a
command that outputs the text of a script to `eval`.