feat: Allow search of `tags` from terminal
This relies on using `ctags` with some specific flags to include the line number
in the created tag entries.

The cryptic `--preview-window` arguments are there to set a fixed 3 line header
for the file name, `~3`, then include that in the scroll offset relative to the
target line number in the second parameter, `+{2}+3, which also includes
centering that line number in the screen by scrolling backwards half a screen,
`,/2`.  Altogether that says `--preview-window="~3,+{2}+3-/2".