feat: Enable LSP
Using a fork of `vim-lsp` because I'm running some LSP servers inside a docker
container, mounting in the PWD.  That works fine except that some of those
servers monitor the supplied editor PID and terminate when that PID dies.
Inside the container they can't see the editor PID so terminate immediately.

The hack is to sent a null editor PID, turning off that feature of the server.