chore: Convert dense refactoring mapping to some functions
This is now an operator i.e. the mapping expects you to have an existing visual
selection, or to follow the mapping with a motion/text object to operate on.

The text to operate on is the value, and the name of the variable is discovered
by searching before an `=` sign.