feat: Add some util vim mappings to extend include search
After an `ilist` or `dlist` these mappings automatically drop you into a prompt
for the particular item to navigate to.  The idea here is that if you're
bothering to look at all the matches you probably want to visit one.

`ijump` and `djump` without an argument go to the first match.