feat: Render images using sixels in iTerm (but not tmux)
`chafa` has the flag `--passthrough=tmux` that should allow sixel rendering from
inside tmux if the underlying terminal emulator supports it, but I could not get
it to work.  I suspect that it's an issue with the `TERM` environment variables
and the `terminfo` database entry that gets picked, but I didn't have time to
get to the bottom of that rabbit hole.