bug: Stop Vim's include-search detecting type definitions
The intent of the original regex, /...(import|:).../, was to teach vim to detect
kotlin imports AND to treat every type mentioned after a `:` as a potential
include that it could look for in the current package (an implicit import).

That works well but unfortunately Vim will not allow the include-search feature
to display import lines... which ruled out any class, function, or interface
definitions that used a `:` to declare a type or declare inheritance.

So I don't think there's a clean way to achieve both a "macro" definition that
includes those core kotlin concepts, and also search for implicit imports in the
current file's package.