feat: Configure vim to preserve views when switching buffers
Vim already preserves the current line when moving between buffers but it does
not preserve the line position, rather placing it in the centre of the screen.
This tweak preserves each buffer's position as ephemeral state on a given window
so the view is only preserved when switching buffers in a single window... which
is really helpful when comparing buffers by switching between them.